refactor: explicit definition of modules
