+++ src/modules/single-course/SingleCourse.module.tsx
