+++ src/pages/course/Course.page.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { SingleCourse } from '../../modules/single-course/SingleCourse';
+import { SingleCourse } from '../../modules/single-course/SingleCourse.module';
 
 export const Course: FC = () => {
   return <SingleCourse />;
