+++ src/pages/home/Home.page.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { CourseList } from '../../modules/course-list/CourseList';
+import { CourseList } from '../../modules/course-list/CourseList.module';
 
 export const Home: FC = () => {
   return <CourseList />;
